requirements.txt
matplotlib~=3.9.2
torch~=2.5.0
opencv-python~=4.10.0.84
imutils~=0.5.4
torchvision~=0.20.0
hydra-core~=1.3.2
numpy~=2.1.2
pillow~=11.0.0
omegaconf~=2.3.0